meetifyr/pyproject.toml
[tool.poetry]
name = "meetifyr"
version = "0.1.0"
description = ""
authors = ["jun2k5 <jun2k5@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqids = "^0.5.2"
cryptography = "44.0.0"
aerich = "0.8.1"
tomlkit = "0.13.2"
pydantic-settings = "2.7.1"
httpx = "0.28.1"

[tool.poetry.dependencies.tortoise-orm]
version = "0.23.0"
extras = ["asyncmy"]

[tool.poetry.group.dec.dependencies.fastapi]
extras = ["standard"]
version = "^0.128.0"

[tool.poetry.group.dev.dependencies]
uvicorn = "^0.40.0"
orjson = "^3.11.5"
black = "^25.12.0"
mypy = "^1.19.1"
coverage = "^7.13.1"
ruff = "^0.14.10"
pytest = "^9.0.2"
pydantic = "^2.12.5"
pytest-asyncio = "^1.3.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"

[tool.coverage.run]
omit = ["*/test_*.py"]

[tool.black]
line-length = 120

[tool.ruff]
target-version = "py314"

[tool.mypy]
plugins = ["pydantic.mypy"]
python_version = 3.13
strict = true

[tool.aerich]
tortoise_orm = "meetifyr.configs.tortoise_config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
